# vityaz-main/openapi.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/api/products/{productId}": {
            "get": {
                "operationId": "ProductsController_getProduct",
                "summary": "Получение определенного товара",
                "parameters": [
                    {
                        "name": "productId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductDto"
                                }
                            }
                        }
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "examples": {
                                    "NotFoundException": {
                                        "description": "Товар не был найден",
                                        "value": {
                                            "statusCode": 404,
                                            "message": "Product not found",
                                            "error": "Not Found"
                                        }
                                    }
                                },
                                "schema": {
                                    "type": "object",
                                    "description": "Товар не был найден",
                                    "properties": {
                                        "statusCode": {
                                            "type": "number",
                                            "example": 404
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Product not found"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Not Found"
                                        }
                                    },
                                    "required": [
                                        "statusCode",
                                        "message"
                                    ]
                                }
                            }
                        },
                        "description": ""
                    }
                },
                "tags": [
                    "Товары"
                ]
            },
            "patch": {
                "operationId": "ProductsController_updateProduct",
                "summary": "Обновление товара (Admin Only)",
                "parameters": [
                    {
                        "name": "productId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer access token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateProductDto"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Товар успешно обновлен"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "examples": {
                                    "UnauthorizedException": {
                                        "description": "Пользователь поставил некорректный токен доступа",
                                        "value": {
                                            "statusCode": 401,
                                            "message": "Invalid access token",
                                            "error": "Unauthorized"
                                        }
                                    }
                                },
                                "schema": {
                                    "type": "object",
                                    "description": "Пользователь поставил некорректный токен доступа",
                                    "properties": {
                                        "statusCode": {
                                            "type": "number",
                                            "example": 401
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid access token"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "required": [
                                        "statusCode",
                                        "message"
                                    ]
                                }
                            }
                        },
                        "description": ""
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "examples": {
                                    "NotFoundException": {
                                        "description": "Товар не был найден",
                                        "value": {
                                            "statusCode": 404,
                                            "message": "Product not found",
                                            "error": "Not Found"
                                        }
                                    }
                                },
                                "schema": {
                                    "type": "object",
                                    "description": "Товар не был найден",
                                    "properties": {
                                        "statusCode": {
                                            "type": "number",
                                            "example": 404
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Product not found"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Not Found"
                                        }
                                    },
                                    "required": [
                                        "statusCode",
                                        "message"
                                    ]
                                }
                            }
                        },
                        "description": ""
                    },
                    "405": {
                        "description": "У пользователя недостаточно прав"
                    }
                },
                "tags": [
                    "Товары",
                    "Для администратора"
                ]
            },
            "delete": {
                "operationId": "ProductsController_deleteProduct",
                "summary": "Удаление товара (Admin Only)",
                "parameters": [
                    {
                        "name": "productId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer access token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Товар успешно удален"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "examples": {
                                    "UnauthorizedException": {
                                        "description": "Пользователь поставил некорректный токен доступа",
                                        "value": {
                                            "statusCode": 401,
                                            "message": "Invalid access token",
                                            "error": "Unauthorized"
                                        }
                                    }
                                },
                                "schema": {
                                    "type": "object",
                                    "description": "Пользователь поставил некорректный токен доступа",
                                    "properties": {
                                        "statusCode": {
                                            "type": "number",
                                            "example": 401
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid access token"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "required": [
                                        "statusCode",
                                        "message"
                                    ]
                                }
                            }
                        },
                        "description": ""
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "examples": {
                                    "NotFoundException": {
                                        "description": "Товар не был найден",
                                        "value": {
                                            "statusCode": 404,
                                            "message": "Product not found",
                                            "error": "Not Found"
                                        }
                                    }
                                },
                                "schema": {
                                    "type": "object",
                                    "description": "Товар не был найден",
                                    "properties": {
                                        "statusCode": {
                                            "type": "number",
                                            "example": 404
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Product not found"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Not Found"
                                        }
                                    },
                                    "required": [
                                        "statusCode",
                                        "message"
                                    ]
                                }
                            }
                        },
                        "description": ""
                    },
                    "405": {
                        "description": "У пользователя недостаточно прав"
                    }
                },
                "tags": [
                    "Товары",
                    "Для администратора"
                ]
            }
        },
...
        "/api/cart": {
            "get": {
                "operationId": "CartsController_getCart",
                "summary": "Получение корзины текущего пользователя",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer access token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Пользователь получает объект корзины",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CartDto"
                                }
                            }
                        }
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "examples": {
                                    "UnauthorizedException": {
                                        "description": "Пользователь поставил некорректный токен доступа",
                                        "value": {
                                            "statusCode": 401,
                                            "message": "Invalid access token",
                                            "error": "Unauthorized"
                                        }
                                    }
                                },
                                "schema": {
                                    "type": "object",
                                    "description": "Пользователь поставил некорректный токен доступа",
                                    "properties": {
                                        "statusCode": {
                                            "type": "number",
                                            "example": 401
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid access token"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "required": [
                                        "statusCode",
                                        "message"
                                    ]
                                }
                            }
                        },
                        "description": ""
                    }
                },
                "tags": [
                    "Корзина"
                ]
            }
        },
...
        "/api/cart/products/{productId}": {
            "post": {
                "operationId": "CartsController_addProductToCart",
                "summary": "Добавление товара в корзину пользователя",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer access token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "productId",
                        "required": true,
                        "in": "path",
                        "description": "Идентификатор товара",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Пользователь успешно добавил товар в свою корзину"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "examples": {
                                    "UnauthorizedException": {
                                        "description": "Пользователь поставил некорректный токен доступа",
                                        "value": {
                                            "statusCode": 401,
                                            "message": "Invalid access token",
                                            "error": "Unauthorized"
                                        }
                                    }
                                },
                                "schema": {
                                    "type": "object",
                                    "description": "Пользователь поставил некорректный токен доступа",
                                    "properties": {
                                        "statusCode": {
                                            "type": "number",
                                            "example": 401
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid access token"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "required": [
                                        "statusCode",
                                        "message"
                                    ]
                                }
                            }
                        },
                        "description": ""
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "examples": {
                                    "NotFoundException": {
                                        "description": "Продукт с указанным идентификатором не был найден",
                                        "value": {
                                            "statusCode": 404,
                                            "message": "Product not found",
                                            "error": "Not Found"
                                        }
                                    }
                                },
                                "schema": {
                                    "type": "object",
                                    "description": "Продукт с указанным идентификатором не был найден",
                                    "properties": {
                                        "statusCode": {
                                            "type": "number",
                                            "example": 404
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Product not found"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Not Found"
                                        }
                                    },
                                    "required": [
                                        "statusCode",
                                        "message"
                                    ]
                                }
                            }
                        },
                        "description": ""
                    }
                },
                "tags": [
                    "Корзина"
                ]
            }
        },
...
    },
    "info": {
        "title": "Магазин «Витязь»",
        "description": "",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "schemas": {
...
            "ProductDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "readOnly": true
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number"
                    },
                    "brand": {
                        "type": "string"
                    },
                    "origin": {
                        "type": "string"
                    },
                    "picture": {
                        "type": "string",
                        "readOnly": true
                    }
                },
                "required": [
                    "id",
                    "name",
                    "price",
                    "brand",
                    "origin",
                    "picture"
                ]
            },
...
            "UpdateProductDto": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number"
                    },
                    "brand": {
                        "type": "string"
                    },
                    "origin": {
                        "type": "string"
                    }
                }
            },
...
            "CartItemDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "description": "Уникальный идентификатор элемента корзины",
                        "nullable": false,
                        "readOnly": true
                    },
                    "product": {
                        "description": "Товар, который пользователь покупает",
                        "readOnly": true,
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/ProductDto"
                            }
                        ]
                    },
                    "quantity": {
                        "type": "number",
                        "description": "Количество конкретно этого товара",
                        "nullable": false,
                        "readOnly": true
                    },
                    "price": {
                        "type": "number",
                        "description": "Цена конкретно этого товара с учетом его количества",
                        "nullable": false,
                        "readOnly": true
                    }
                },
                "required": [
                    "id",
                    "product",
                    "quantity",
                    "price"
                ]
            },
            "CartDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "description": "Уникальный идентификатор корзины пользователя",
                        "nullable": false,
                        "readOnly": true
                    },
                    "items": {
                        "description": "Элементы корзины",
                        "nullable": false,
                        "readOnly": true,
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CartItemDto"
                        }
                    },
                    "totalPrice": {
                        "type": "number",
                        "description": "Общая стоимость товаров в корзине",
                        "nullable": false,
                        "readOnly": true
                    }
                },
                "required": [
                    "id",
                    "items",
                    "totalPrice"
                ]
            },
...
        }
    }
}